+++ backend/database/favourite_db.js
@@ -6,7 +6,10 @@ const sql = {
 }
 
 async function addFavourite(idaccount, mdbdata) {
-    let result = await pgPool.query(sql.ADD_FAVOURITE, [idaccount, mdbdata]);
+    // Convert mdbdata to a string representation
+    const truncatedMdbdata = JSON.stringify(mdbdata).substring(0, 255);
+    
+    let result = await pgPool.query(sql.ADD_FAVOURITE, [idaccount, truncatedMdbdata]);
     return result.rows;
 }
