+++ frontend/src/components/addfavouriteicon.css
@@ -1,0 +1,6 @@
+.addfavouriteicon{
+    font-size: 300%;
+    margin: 2%;
+    color: rgb(214, 82, 104);
+    cursor: pointer;
+}
